phase-4/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "todo-backend-api"
version = "0.1.0"
description = "Backend API for Todo application with user ownership"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2,bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "sqlmodel>=0.0.16",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.7",
    "alembic>=1.13.0",
    "sqlalchemy>=2.0.0",
    "openai-agents>=0.6.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "ruff>=0.1.6",
    "black>=23.0",
    "mypy>=1.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = [
    "E",  # pycodestyle
    "F",  # Pyflakes
    "I",  # isort
    "PL", # pylint (subset)
]
line-length = 88
extend-ignore = [
    "PLR",  # Design related pylint codes
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
